# improve performance of safeParse