@Uzlopak 2022-07-29
@mcollina 

Should not be backwards breaking. I just corrected the typings of scan.