@mcollina 2022-07-29
Can you send a PR to fastify to use this instead? It should improve things in case of a failure.